+++ CHANGELOG.md
@@ -5,7 +5,12 @@ We are currently working on porting this changelog to the specifications in
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Version 0.7.0 - Unreleased
+## Version 0.7.1 - Unreleased
+
+### Fixed:
+* Fixed issue where imdelay would sometimes not respect antialias and interpolation
+
+## Version 0.7.0 - Released 2023-08-15
 
 ### Added